requirements.txt
# Core Framework
langchain
langchain-openai
langchain-community
langgraph

# RAG & PDF Processing
pypdf
faiss-cpu
beautifulsoup4

# Web & Environment
fastapi
uvicorn
python-dotenv
pydantic>=2.0